requirements.txt
pyspark==3.3.0
pytest==7.4.0
pytest-cov==4.1.0
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.1.0
psycopg2-binary==2.9.9
cryptography==41.0.7 